refactor: rename type to types and auth to isAuthorized
